requirements.txt
discord.py
cryptography
requests
aiohttp
assemblyai
dotenv
yt-dlp